+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 
 - rotation angle for Rect
 - rotate method for Point
+- Point.from now accepts PointArray
 
 ## [0.0.6] - 2023-05-30